FIX: sign-unsign comparison warning: loop index variable type, gtest ASSERT and EXPECT comparisons.

+++ src/common/test/test_padding.cpp
@@ -18,8 +18,8 @@ TEST(PaddingTest, RemovePaddingSameType)
   uint32_t* begin = paddedseq;
   bliss::common::removePadding(begin, unpadded, 4*16, 16);
 
-  EXPECT_EQ(unpadded[0], 0xbeefdead) << std::hex << "wrong result is: 0x"  << unpadded[0];
-  EXPECT_EQ(unpadded[1], 0xabba1234) << std::hex << "wrong result is: 0x"  << unpadded[1];
+  EXPECT_EQ(unpadded[0], static_cast<uint32_t>(0xbeefdead)) << std::hex << "wrong result is: 0x"  << unpadded[0];
+  EXPECT_EQ(unpadded[1], static_cast<uint32_t>(0xabba1234)) << std::hex << "wrong result is: 0x"  << unpadded[1];
 }
 
 TEST(PaddingTest, RemovePaddingSameType2)
@@ -31,11 +31,11 @@ TEST(PaddingTest, RemovePaddingSameType2)
   bliss::common::removePadding(begin, unpadded, 6*6*4, 8);
 
   // check for correct depadding (entered manually)
-  EXPECT_EQ(unpadded[0], 0xef12dead) << std::hex << "wrong result is: 0x"  << unpadded[0];
-  EXPECT_EQ(unpadded[1], 0x123434be) << std::hex << "wrong result is: 0x"  << unpadded[1];
-  EXPECT_EQ(unpadded[2], 0x56abbaff) << std::hex << "wrong result is: 0x"  << unpadded[2];
-  EXPECT_EQ(unpadded[3], 0xffc0ffee) << std::hex << "wrong result is: 0x"  << unpadded[3];
-  EXPECT_EQ(unpadded[4], 0x0000deca) << std::hex << "wrong result is: 0x"  << unpadded[4];
+  EXPECT_EQ(unpadded[0], static_cast<uint32_t>(0xef12dead)) << std::hex << "wrong result is: 0x"  << unpadded[0];
+  EXPECT_EQ(unpadded[1], static_cast<uint32_t>(0x123434be)) << std::hex << "wrong result is: 0x"  << unpadded[1];
+  EXPECT_EQ(unpadded[2], static_cast<uint32_t>(0x56abbaff)) << std::hex << "wrong result is: 0x"  << unpadded[2];
+  EXPECT_EQ(unpadded[3], static_cast<uint32_t>(0xffc0ffee)) << std::hex << "wrong result is: 0x"  << unpadded[3];
+  EXPECT_EQ(unpadded[4], static_cast<uint32_t>(0x0000deca)) << std::hex << "wrong result is: 0x"  << unpadded[4];
 }
 
 
@@ -49,8 +49,8 @@ TEST(PaddingTest, RemovePaddingSameType3)
   bliss::common::removePadding(begin, unpadded, 4*13, 32-13);
 
   // check for correct depadding (entered manually)
-  EXPECT_EQ(unpadded[0], 0xd3ddfead) << std::hex << "wrong result is: 0x"  << unpadded[0];
-  EXPECT_EQ(unpadded[1], 0x0005dd48) << std::hex << "wrong result is: 0x"  << unpadded[1];
+  EXPECT_EQ(unpadded[0], static_cast<uint32_t>(0xd3ddfead)) << std::hex << "wrong result is: 0x"  << unpadded[0];
+  EXPECT_EQ(unpadded[1], static_cast<uint32_t>(0x0005dd48)) << std::hex << "wrong result is: 0x"  << unpadded[1];
 }
 
 TEST(PaddingTest, RemovePadding32to16by8)
@@ -62,15 +62,15 @@ TEST(PaddingTest, RemovePadding32to16by8)
   bliss::common::removePadding(begin, unpadded, 6*6*4, 8);
 
   // check for correct depadding (entered manually)
-  EXPECT_EQ(unpadded[0], 0xdead) << std::hex << "wrong result is: 0x"  << unpadded[0];
-  EXPECT_EQ(unpadded[1], 0xef12) << std::hex << "wrong result is: 0x"  << unpadded[1];
-  EXPECT_EQ(unpadded[2], 0x34be) << std::hex << "wrong result is: 0x"  << unpadded[2];
-  EXPECT_EQ(unpadded[3], 0x1234) << std::hex << "wrong result is: 0x"  << unpadded[3];
-  EXPECT_EQ(unpadded[4], 0xbaff) << std::hex << "wrong result is: 0x"  << unpadded[4];
-  EXPECT_EQ(unpadded[5], 0x56ab) << std::hex << "wrong result is: 0x"  << unpadded[5];
-  EXPECT_EQ(unpadded[6], 0xffee) << std::hex << "wrong result is: 0x"  << unpadded[6];
-  EXPECT_EQ(unpadded[7], 0xffc0) << std::hex << "wrong result is: 0x"  << unpadded[7];
-  EXPECT_EQ(unpadded[8], 0xdeca) << std::hex << "wrong result is: 0x"  << unpadded[8];
+  EXPECT_EQ(unpadded[0], static_cast<uint16_t>(0xdead)) << std::hex << "wrong result is: 0x"  << unpadded[0];
+  EXPECT_EQ(unpadded[1], static_cast<uint16_t>(0xef12)) << std::hex << "wrong result is: 0x"  << unpadded[1];
+  EXPECT_EQ(unpadded[2], static_cast<uint16_t>(0x34be)) << std::hex << "wrong result is: 0x"  << unpadded[2];
+  EXPECT_EQ(unpadded[3], static_cast<uint16_t>(0x1234)) << std::hex << "wrong result is: 0x"  << unpadded[3];
+  EXPECT_EQ(unpadded[4], static_cast<uint16_t>(0xbaff)) << std::hex << "wrong result is: 0x"  << unpadded[4];
+  EXPECT_EQ(unpadded[5], static_cast<uint16_t>(0x56ab)) << std::hex << "wrong result is: 0x"  << unpadded[5];
+  EXPECT_EQ(unpadded[6], static_cast<uint16_t>(0xffee)) << std::hex << "wrong result is: 0x"  << unpadded[6];
+  EXPECT_EQ(unpadded[7], static_cast<uint16_t>(0xffc0)) << std::hex << "wrong result is: 0x"  << unpadded[7];
+  EXPECT_EQ(unpadded[8], static_cast<uint16_t>(0xdeca)) << std::hex << "wrong result is: 0x"  << unpadded[8];
 }
 
 TEST(PaddingTest, RemovePadding32to8by28)
@@ -82,9 +82,9 @@ TEST(PaddingTest, RemovePadding32to8by28)
   bliss::common::removePadding(begin, unpadded, 4*6, 28);
 
   // check for correct depadding (entered manually)
-  EXPECT_EQ(unpadded[0], 0xfd) << std::hex << "wrong result is: 0x"  << unpadded[0];
-  EXPECT_EQ(unpadded[1], 0xa4) << std::hex << "wrong result is: 0x"  << unpadded[1];
-  EXPECT_EQ(unpadded[2], 0xfe) << std::hex << "wrong result is: 0x"  << unpadded[2];
+  EXPECT_EQ(unpadded[0], static_cast<uint8_t>(0xfd)) << std::hex << "wrong result is: 0x"  << unpadded[0];
+  EXPECT_EQ(unpadded[1], static_cast<uint8_t>(0xa4)) << std::hex << "wrong result is: 0x"  << unpadded[1];
+  EXPECT_EQ(unpadded[2], static_cast<uint8_t>(0xfe)) << std::hex << "wrong result is: 0x"  << unpadded[2];
 }
 
 TEST(PaddingTest, RemovePadding32to16by13)
@@ -96,10 +96,10 @@ TEST(PaddingTest, RemovePadding32to16by13)
   bliss::common::removePadding(begin, unpadded, 4*13, 32-13);
 
   // check for correct depadding (entered manually)
-  EXPECT_EQ(unpadded[0], 0xfead) << std::hex << "wrong result is: 0x"  << unpadded[0];
-  EXPECT_EQ(unpadded[1], 0xd3dd) << std::hex << "wrong result is: 0x"  << unpadded[1];
-  EXPECT_EQ(unpadded[2], 0xdd48) << std::hex << "wrong result is: 0x"  << unpadded[2];
-  EXPECT_EQ(unpadded[3], 0x0005) << std::hex << "wrong result is: 0x"  << unpadded[3];
+  EXPECT_EQ(unpadded[0], static_cast<uint16_t>(0xfead)) << std::hex << "wrong result is: 0x"  << unpadded[0];
+  EXPECT_EQ(unpadded[1], static_cast<uint16_t>(0xd3dd)) << std::hex << "wrong result is: 0x"  << unpadded[1];
+  EXPECT_EQ(unpadded[2], static_cast<uint16_t>(0xdd48)) << std::hex << "wrong result is: 0x"  << unpadded[2];
+  EXPECT_EQ(unpadded[3], static_cast<uint16_t>(0x0005)) << std::hex << "wrong result is: 0x"  << unpadded[3];
 }
 
 TEST(PaddingTest, RemovePadding32to8by13)
@@ -111,11 +111,11 @@ TEST(PaddingTest, RemovePadding32to8by13)
   bliss::common::removePadding(begin, unpadded, 4*13, 32-13);
 
   // check for correct depadding (entered manually)
-  EXPECT_EQ(unpadded[0], 0xad) << std::hex << "wrong result is: 0x"  << unpadded[0];
-  EXPECT_EQ(unpadded[1], 0xfe) << std::hex << "wrong result is: 0x"  << unpadded[1];
-  EXPECT_EQ(unpadded[2], 0xdd) << std::hex << "wrong result is: 0x"  << unpadded[2];
-  EXPECT_EQ(unpadded[3], 0xd3) << std::hex << "wrong result is: 0x"  << unpadded[3];
-  EXPECT_EQ(unpadded[4], 0x48) << std::hex << "wrong result is: 0x"  << unpadded[4];
-  EXPECT_EQ(unpadded[5], 0xdd) << std::hex << "wrong result is: 0x"  << unpadded[5];
-  EXPECT_EQ(unpadded[6], 0x05) << std::hex << "wrong result is: 0x"  << unpadded[6];
+  EXPECT_EQ(unpadded[0], static_cast<uint8_t>(0xad)) << std::hex << "wrong result is: 0x"  << unpadded[0];
+  EXPECT_EQ(unpadded[1], static_cast<uint8_t>(0xfe)) << std::hex << "wrong result is: 0x"  << unpadded[1];
+  EXPECT_EQ(unpadded[2], static_cast<uint8_t>(0xdd)) << std::hex << "wrong result is: 0x"  << unpadded[2];
+  EXPECT_EQ(unpadded[3], static_cast<uint8_t>(0xd3)) << std::hex << "wrong result is: 0x"  << unpadded[3];
+  EXPECT_EQ(unpadded[4], static_cast<uint8_t>(0x48)) << std::hex << "wrong result is: 0x"  << unpadded[4];
+  EXPECT_EQ(unpadded[5], static_cast<uint8_t>(0xdd)) << std::hex << "wrong result is: 0x"  << unpadded[5];
+  EXPECT_EQ(unpadded[6], static_cast<uint8_t>(0x05)) << std::hex << "wrong result is: 0x"  << unpadded[6];
 }
